Struct secp256k1::Secp256k1

pub struct Secp256k1<C: Context> { /* fields omitted */ }

The secp256k1 engine, used to execute all signature operations

Methods

impl<C: Context> Secp256k1<C>

pub fn gen_new() -> Secp256k1<C>

Lets you create a context in a generic manner(sign/verify/all)

impl Secp256k1<All>

pub fn new() -> Secp256k1<All>

Creates a new Secp256k1 context with all capabilities

impl Secp256k1<SignOnly>

pub fn signing_only() -> Secp256k1<SignOnly>

Creates a new Secp256k1 context that can only be used for signing

impl Secp256k1<VerifyOnly>

pub fn verification_only() -> Secp256k1<VerifyOnly>

Creates a new Secp256k1 context that can only be used for verification

impl<'buf, C: Context + 'buf> Secp256k1<C>

pub fn preallocated_gen_new(buf: &'buf mut [u8]) -> Result<Secp256k1<C>, Error>

Lets you create a context with preallocated buffer in a generic manner(sign/verify/all)

impl<'buf> Secp256k1<AllPreallocated<'buf>>

pub fn preallocated_new(
buf: &'buf mut [u8]
) -> Result<Secp256k1<AllPreallocated<'buf>>, Error>

Creates a new Secp256k1 context with all capabilities

pub fn preallocate_size() -> usize

Uses the ffi secp256k1_context_preallocated_size to check the memory size needed for a context

impl<'buf> Secp256k1<SignOnlyPreallocated<'buf>>

pub fn preallocated_signing_only(
buf: &'buf mut [u8]
) -> Result<Secp256k1<SignOnlyPreallocated<'buf>>, Error>

Creates a new Secp256k1 context that can only be used for signing

pub fn preallocate_signing_size() -> usize

Uses the ffi secp256k1_context_preallocated_size to check the memory size needed for the context

impl<'buf> Secp256k1<VerifyOnlyPreallocated<'buf>>

pub fn preallocated_verification_only(
buf: &'buf mut [u8]
) -> Result<Secp256k1<VerifyOnlyPreallocated<'buf>>, Error>

Creates a new Secp256k1 context that can only be used for verification

pub fn preallocate_verification_size() -> usize

Uses the ffi secp256k1_context_preallocated_size to check the memory size needed for the context

impl<C: Context> Secp256k1<C>

pub fn ctx(&self) -> &*mut Context

Getter for the raw pointer to the underlying secp256k1 context. This shouldn't be needed with normal usage of the library. It enables extending the Secp256k1 with more cryptographic algorithms outside of this crate.

pub fn preallocate_size_gen() -> usize

Returns the required memory for a preallocated context buffer in a generic manner(sign/verify/all)

impl<C: Signing> Secp256k1<C>

pub fn sign(&self, msg: &Message, sk: &SecretKey) -> Signature

Constructs a signature for msg using the secret key sk and RFC6979 nonce Requires a signing-capable context.

impl<C: Verification> Secp256k1<C>

pub fn verify(
&self,
msg: &Message,
sig: &Signature,
pk: &PublicKey
) -> Result<(), Error>

Checks that sig is a valid ECDSA signature for msg using the public key pubkey. Returns Ok(true) on success. Note that this function cannot be used for Bitcoin consensus checking since there may exist signatures which OpenSSL would verify but not libsecp256k1, or vice-versa. Requires a verify-capable context.